Rename peer_selection to select
